Ensure that Calypso always displays a hierarchical view for pages
Currently, we display the hierarchical view for the site that has up
to 100 pages. However, with an infinite scroll in place, the user sees
a chronological view until all records all loaded, and then the view
switches to hierarchical.

I try to improve this experience by loading 100 pages at once,
to trigger the hierarchical view immediately.